backend/requirements.txt
fastapi>=0.109
uvicorn>=0.27
sqlalchemy>=2.0
alembic>=1.13
pydantic>=2.5
pydantic-settings>=2.1
python-jose>=3.3
passlib>=1.7
python-multipart>=0.0.6
python-dotenv>=1.0
requests>=2.32
httpx>=0.27
vercel
